requirements.txt
streamlit
streamlit_extras
streamlit_pdf_viewer
google.generativeai
langchain_text_splitters
langchain_community
bump-my-version
